+++ BANPAIS_SOAP_API/src/main/java/com/banpais/api/service/CuentaSoapService.java
@@ -17,6 +17,7 @@ import java.util.Map;
 import java.math.BigDecimal;
 import java.math.RoundingMode;
 import java.util.UUID;
+import org.springframework.beans.factory.annotation.Value;
 
 @Service
 public class CuentaSoapService {
@@ -25,21 +26,21 @@ public class CuentaSoapService {
     private final ClienteRepository clienteRepository;
     private final TramaService tramaService;
     private static final String CUENTA_ENTRADA = "CUENTA_IN";
+    @Value("${banco.cuenta.oficial.numero}")
+    private String bancoCuentaOficial;
 
     // Constantes para los nombres de los campos
     private static final String CAMPO_NUMERO_CUENTA = "NUMERO_CUENTA";
     private static final String CAMPO_FK_ID_CLIENTE = "FK_ID_CLIENTE";
     private static final String CAMPO_FECHA_APERTURA = "FECHA_APERTURA";
-    private static final String CAMPO_ESTADO_CUENTA = "ESTADO_CUENTA"; // Aunque no se modifique directamente, es bueno tenerlo.
+    private static final String CAMPO_ESTADO_CUENTA = "ESTADO_CUENTA";
     private static final String CAMPO_SALDO = "SALDO";
 
-    //Considera usar un enum para el estado
     public enum EstadoCuenta {
-        ACT, 
-        INA, 
-        BLO  
+        ACT,
+        INA,
+        BLO
     }
-
 
     public CuentaSoapService(CuentaRepository cuentaRepository, ClienteRepository clienteRepository, TramaService tramaService) {
         this.cuentaRepository = cuentaRepository;
@@ -60,7 +61,6 @@ public class CuentaSoapService {
         System.out.println("trama plana" + trama);
         System.out.println("campos" + campos.toString());
 
-
         Cliente cliente = clienteRepository.findById(campos.get(CAMPO_FK_ID_CLIENTE))
                 .orElseThrow(() -> new ClienteNotFoundException("No se encontró el cliente con el ID: " + campos.get(CAMPO_FK_ID_CLIENTE)));
 
@@ -68,7 +68,6 @@ public class CuentaSoapService {
         cuenta.setNumeroCuenta(UUID.randomUUID().toString().replace("-", "").substring(0, 16));
         cuenta.setCliente(campos.get(CAMPO_FK_ID_CLIENTE));
         cuenta.setFechaApertura(tramaService.parseLocalDateTime(campos.get(CAMPO_FECHA_APERTURA)));
-
 
         LocalDateTime fechaHoraApertura = tramaService.parseLocalDateTime(campos.get(CAMPO_FECHA_APERTURA));
         cuenta.setHoraApertura(fechaHoraApertura != null ? fechaHoraApertura.toLocalTime() : null);
@@ -81,8 +80,8 @@ public class CuentaSoapService {
         cuentaRepository.save(cuenta);
 
         RegistrarCuentaResponse response = new RegistrarCuentaResponse();
-        response.setCodigo("000"); 
-        response.setMensaje("Cuenta creada exitosamente"); 
+        response.setCodigo("000");
+        response.setMensaje("Cuenta creada exitosamente");
         return response;
     }
 
@@ -114,8 +113,8 @@ public class CuentaSoapService {
         cuentaRepository.save(cuenta);
 
         ActualizarCuentaResponse response = new ActualizarCuentaResponse();
-        response.setCodigo("000"); 
-        response.setMensaje("Cuenta actualizada exitosamente"); 
+        response.setCodigo("000");
+        response.setMensaje("Cuenta actualizada exitosamente");
         return response;
     }
 
@@ -123,15 +122,21 @@ public class CuentaSoapService {
     public EliminarCuentaResponse eliminarCuenta(String trama) {
         List<TramaParametro> parametros = tramaService.getParametrosTrama(CUENTA_ENTRADA);
         Map<String, String> campos = tramaService.desparametrizarTrama(trama, parametros);
+        
+        String numeroCuenta = campos.get(CAMPO_NUMERO_CUENTA);
+        System.out.println("comparativa "+numeroCuenta+bancoCuentaOficial);
+        if (numeroCuenta.equals(bancoCuentaOficial)) {
+            throw new RuntimeException("No se puede eliminar la cuenta oficial del banco");
+        }
 
-        Cuenta cuenta = cuentaRepository.findById(campos.get(CAMPO_NUMERO_CUENTA))
-                .orElseThrow(() -> new CuentaNotFoundException("No existe la cuenta con el número de cuenta: " + campos.get(CAMPO_NUMERO_CUENTA)));
+        Cuenta cuenta = cuentaRepository.findById(numeroCuenta)
+                .orElseThrow(() -> new CuentaNotFoundException("No existe la cuenta con el número de cuenta: " + numeroCuenta));
 
         cuentaRepository.delete(cuenta);
 
         EliminarCuentaResponse response = new EliminarCuentaResponse();
-        response.setCodigo("000"); 
-        response.setMensaje("Cuenta eliminada exitosamente"); 
+        response.setCodigo("000");
+        response.setMensaje("Cuenta eliminada exitosamente");
         return response;
     }
 }
